isometric

+++ src/matrix/center_mtx.c
@@ -5,11 +5,11 @@ void	center_obj_matrix(t_mlx *param, double matrix[4][4])
 	matrix[0][0] = 1;
 	matrix[0][1] = 0;
 	matrix[0][2] = 0;
-	matrix[0][3] = -(param->x_max / 2);
+	matrix[0][3] = -((double) param->x_max / 2);
 	matrix[1][0] = 0;
 	matrix[1][1] = 1;
 	matrix[1][2] = 0;
-	matrix[1][3] = -(param->y_max / 2);
+	matrix[1][3] = -((double) param->y_max / 2);
 	matrix[2][0] = 0;
 	matrix[2][1] = 0;
 	matrix[2][2] = 1;
